requirements.txt
streamlit>=1.33.0
pandas>=2.2.0
scikit-learn>=1.4.0
plotly-express>=0.4.1
seaborn>=0.13.0